QOL improvements and pagination fix into 25-1
Big improvement is that it should output summaries by assignment group, so in this example at the end is the summary of the selected labs and WPRs.

Other improvements:
I fixed the app crashing on first run each time.
I fixed it so you can select more than 5 courses in the data pull.
I changed the course selection to be a table that is filterable. This was necessary since you will have a lot of courses on your account now.
With that you can still save default courses to save some time selecting them all
This table will also let you filter by section and hour
I added the assignment group summarizing to help with OML.